(FFM-5326) Use go evaluator for getfeature endpoint
