+++ go.mod
@@ -11,7 +11,7 @@ require (
 	github.com/golang-jwt/jwt v3.2.2+incompatible
 	github.com/golang-jwt/jwt/v4 v4.4.2
 	github.com/google/uuid v1.3.0
-	github.com/harness/ff-golang-server-sdk v0.1.3
+	github.com/harness/ff-golang-server-sdk v0.1.5
 	github.com/hashicorp/go-retryablehttp v0.6.8
 	github.com/joho/godotenv v1.4.0
 	github.com/labstack/echo/v4 v4.9.0
